Moved Span3D actions to their own file
