+++ app/src/components/toolbar.tsx
@@ -12,10 +12,9 @@ import {
   changeViewerConfig,
   mergeTracks,
   splitTrack,
-  startLinkTrack,
-  activateSpan,
-  deactivateSpan
+  startLinkTrack
 } from "../action/common"
+import { activateSpan, deactivateSpan } from "../action/span3d"
 import {
   changeSelectedLabelsAttributes,
   deleteSelectedLabels,
@@ -25,7 +24,7 @@ import {
 import { addLabelTag } from "../action/tag"
 import { renderTemplate } from "../common/label"
 import Session from "../common/session"
-import { Key, LabelTypeName } from "../const/common"
+import { ItemTypeName, Key, LabelTypeName } from "../const/common"
 import { getSelectedTracks } from "../functional/state_util"
 import { isValidId, makeTrack } from "../functional/states"
 import { tracksOverlapping } from "../functional/track"
@@ -187,16 +186,20 @@ export class ToolBar extends Component<Props> {
               this.deletePressed()
             })}
           </div>
-          <div>
-            {this.state.session.boxSpan || this.state.task.boxSpan !== undefined
-              ? makeButton("Cancel", () => {
-                  this.deactivateSpan()
-                  alert(Severity.WARNING, "Box was not generated")
-                })
-              : makeButton("Activate span", () => {
-                  this.activateSpan()
-                })}
-          </div>
+          {(this.state.task.config.itemType === ItemTypeName.POINT_CLOUD ||
+            this.state.task.config.itemType ===
+              ItemTypeName.POINT_CLOUD_TRACKING) && (
+            <div>
+              {this.state.session.boxSpan || this.state.task.boxSpan !== null
+                ? makeButton("Cancel", () => {
+                    this.deactivateSpan()
+                    alert(Severity.WARNING, "Box was not generated")
+                  })
+                : makeButton("Activate span", () => {
+                    this.activateSpan()
+                  })}
+            </div>
+          )}
           {this.state.task.config.tracking && (
             <div>
               {this.state.session.trackLinking
