+++ app/src/drawable/3d/box_span/spanPoint3d.ts
@@ -6,16 +6,13 @@ import { Vector3D } from "../../../math/vector3d"
  * ThreeJS class for rendering 3D point
  */
 export class SpanPoint3D {
+  private readonly _point: THREE.Mesh
   /** x coordinate in 3D space */
   private readonly _x: number
   /** y coordinate in 3D space */
   private readonly _y: number
   /** z coordinate in 3D space */
   private readonly _z: number
-  /** point color */
-  private readonly _color: string
-  /** point radius */
-  private readonly _radius: number
 
   /**
    * Constructor
@@ -26,8 +23,9 @@ export class SpanPoint3D {
     this._x = point.x
     this._y = point.y
     this._z = point.z
-    this._color = "#ff0000"
-    this._radius = 0.05
+    const geometry = new THREE.SphereGeometry(0.05)
+    const material = new THREE.MeshBasicMaterial({ color: "#ff0000" })
+    this._point = new THREE.Mesh(geometry, material)
   }
 
   /**
@@ -38,11 +36,8 @@ export class SpanPoint3D {
    * @param pointCloud
    */
   public render(scene: THREE.Scene): void {
-    const geometry = new THREE.SphereGeometry(this._radius)
-    const material = new THREE.MeshBasicMaterial({ color: this._color })
-    const point = new THREE.Mesh(geometry, material)
-    point.position.set(this.x, this.y, this.z)
-    scene.add(point)
+    this._point.position.set(this.x, this.y, this.z)
+    scene.add(this._point)
   }
 
   /** get x */
